Tried antiquating the stod

+++ main.cpp
@@ -34,9 +34,11 @@ void handle_walk(const std::string & r){
             c2 += r[curr2];
             curr2++;
         } else{
-            speed = std::stod(c2);
+            //speed = std::stod(c2);
+            speed = std::atoi(c2);
             r2 = r.substr(curr2 + 1, r.length());
-            angle = std::stod(r2);
+            //angle = std::stod(r2);
+            angle = std::atoi(r2);
             controller->move(speed, angle);
             break;
         }
